requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
pytest==7.4.3
pytest-flask==1.3.0
